+++ .github/CHANGELOG.md
@@ -27,6 +27,39 @@
   [(#819)](https://github.com/PennyLaneAI/pennylane/pull/819)
   [(#807)](https://github.com/PennyLaneAI/pennylane/pull/807)
   [(#794)](https://github.com/PennyLaneAI/pennylane/pull/794)
+=======
+* The quantum state of a QNode can now be returned using the ``state()`` return function.
+  [(#818)](https://github.com/XanaduAI/pennylane/pull/818)
+  
+  Consider the following QNode:
+  ```python
+  import pennylane as qml
+  from pennylane.beta.tapes import qnode
+  from pennylane.beta.queuing import state
+  
+  dev = qml.device("default.qubit", wires=3)
+  
+  @qnode(dev)
+  def qfunc(x, y):
+      qml.RZ(x, wires=0)
+      qml.CNOT(wires=[0, 1])
+      qml.RY(y, wires=1)
+      qml.CNOT(wires=[0, 2])
+      return state()
+  ```
+  
+  Calling the QNode will return its state
+  
+  ```pycon
+  >>> qfunc(0.56, 0.1)
+  array([0.95985437-0.27601028j, 0.        +0.j        ,
+       0.04803275-0.01381203j, 0.        +0.j        ,
+       0.        +0.j        , 0.        +0.j        ,
+       0.        +0.j        , 0.        +0.j        ])
+  ```
+  
+  Differentiating the state is not yet fully supported, but is currently available when using the
+  classical backpropagation differentiation method (``diff_method="backprop"``) with a compatible device.
 
 * Summation of two `Wires` objects is now supported and will return 
   a `Wires` object containing the set of all wires defined by the 
@@ -52,6 +85,29 @@
 
 
 <h3>Improvements</h3>
+
+* QNode caching has been introduced, allowing the QNode to keep track of the results of previous
+  device executions and reuse those results in subsequent calls.
+  [(#817)](https://github.com/PennyLaneAI/pennylane/pull/817)
+
+  Caching is available by passing a ``caching`` argument to the QNode:
+  
+  ```python
+  from pennylane.beta.tapes import qnode
+  from pennylane.beta.queuing import expval
+    
+  dev = qml.device("default.qubit", wires=2)
+    
+  @qnode(dev, caching=10)  # cache up to 10 evaluations
+  def qfunc(x):
+      qml.RX(x, wires=0)
+      qml.RX(0.3, wires=1)
+      qml.CNOT(wires=[0, 1])
+      return expval(qml.PauliZ(1))
+    
+  qfunc(0.1)  # first evaluation executes on the device
+  qfunc(0.1)  # second evaluation accesses the cached result
+  ```
 
 * Sped up the application of certain gates in `default.qubit` by using array/tensor
   manipulation tricks. The following gates are affected: `PauliX`, `PauliY`, `PauliZ`,